refactor the routes
